Implement Stellar Provider Availability Heatmap
